+++ compiler/fhir-parser/src/loader.rs
@@ -1,0 +1,7 @@
+use fhir_ir::Specification;
+
+use crate::LoaderError;
+
+pub trait Loader {
+    fn load(&self) -> Result<Specification, LoaderError>;
+}